refactor: move curp test utils to `curp-test-utils` crate
This commit move both unit and integration test utils to a seperated crate, which can be reused.

Closes #290
