+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 
-members = ["xline", "curp", "benchmark", "utils", "engine", "xlineapi", "curp-external-api"]
+members = ["xline", "curp", "benchmark", "utils", "engine", "xlineapi", "curp-external-api", "curp-test-utils"]
