+++ curp-test-utils/README.md
@@ -1,0 +1,3 @@
+# Curp test utilities
+
+This crate provides utilities for curp tests.
