+++ curp/Cargo.toml
@@ -45,6 +45,7 @@ uuid = { version = "1.3.1", features = ["v4"] }
 bytes = "1.4.0"
 
 [dev-dependencies]
+curp-test-utils = { path = "../curp-test-utils" }
 itertools = "0.10.3"
 tracing-subscriber = { version = "0.3.16", features = ["env-filter", "time"] }
 tracing-test = "0.2.4"
